+++ uniffi_bindgen/src/bindings/kotlin/gen_kotlin/primitives.rs
@@ -67,6 +67,7 @@ macro_rules! impl_code_type_for_primitive {
 
 impl_code_type_for_primitive!(BooleanCodeType, "Boolean");
 impl_code_type_for_primitive!(StringCodeType, "String");
+impl_code_type_for_primitive!(BytesCodeType, "ByteArray");
 impl_code_type_for_primitive!(Int8CodeType, "Byte");
 impl_code_type_for_primitive!(Int16CodeType, "Short");
 impl_code_type_for_primitive!(Int32CodeType, "Int");