+++ uniffi_bindgen/src/bindings/kotlin/templates/Types.kt
@@ -53,6 +53,9 @@
 {%- when Type::String %}
 {%- include "StringHelper.kt" %}
 
+{%- when Type::Bytes %}
+{%- include "ByteArrayHelper.kt" %}
+
 {%- when Type::Enum(name) %}
 {% include "EnumTemplate.kt" %}