+++ uniffi_bindgen/src/bindings/python/gen_python/primitives.rs
@@ -57,6 +57,7 @@ macro_rules! impl_code_type_for_primitive {
 
 impl_code_type_for_primitive!(BooleanCodeType, "Bool", "bool({})");
 impl_code_type_for_primitive!(StringCodeType, "String", "{}");
+impl_code_type_for_primitive!(BytesCodeType, "Bytes", "{}");
 impl_code_type_for_primitive!(Int8CodeType, "Int8", "int({})");
 impl_code_type_for_primitive!(Int16CodeType, "Int16", "int({})");
 impl_code_type_for_primitive!(Int32CodeType, "Int32", "int({})");